requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cors==1.0.1
dnspython==2.3.0
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.2
requests-file==1.5.1
six==1.16.0
tldextract==3.4.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.13.0
zope.event==4.6
zope.interface==5.5.2
